+++ rai/models.go
@@ -411,6 +411,16 @@ type createSnowflakeDatabaseLinkRequest struct {
 	} `json:"snowflake"`
 }
 
+type updateSnowflakeDatabaseLinkRequest struct {
+	Snowflake struct {
+		Database    string               `json:"database"`
+		Schema      string               `json:"schema"`
+		Role        string               `json:"role"`
+		Warehouse   string               `json:"warehouse"`
+		Credentials SnowflakeCredentials `json:"credentials"` // not-persisted
+	} `json:"snowflake"`
+}
+
 type deleteSnowflakeDatabaseLinkRequest struct {
 	Snowflake struct {
 		Role        string               `json:"role"`